Don't insert "No changes" for a pre-existing version header
This makes `Changelog::Updater` idempotent. If we perform a release,
updating the changelog and removing the data files, then abort for some
reason during a later phase, re-running the release process will no
longer insert "No changes" directly above a bunch of changes that we
just inserted during the previous run.

To facilitate this, `Changelog::Updater#insert` no longer takes a String
of Markdown to insert, but instead takes the `MarkdownGenerator` object
so that we can query its `entries` property before performing the
insert.