Make sure we merge with `--no-log`
I realized that the test keeps failing for me,
and that's because I set `git config merge.log true`
so it's generating a different commit message.
(with updated files list)

With `--no-log` explicitly set, the test passed for me
even when it's set true on my computer. This should
make sure we have consistent result.