Adds new rake tasks for the purpose of generating QA issues
Task 1: FirstRcQaIssue

Generates a list of closed feature proposals and bugs for the purpose
of feature assurance and quality assurance of the first release candidate
of a new release.

This list should encompass all bug fixes and new features.

Task 2: RcPatchQaIssue

Generate a list of merged merge requests for the purpose of quality
assurance of the latest Release Candidate or Patch release.

This list should encompass all bug fixes merged into the `stable`
branch for the current release.