Create an issue indexing changes between two refs
Feature introduced for the purpose of generating a QA task per version.

Current QA strategy entails creating one issue to track manual testing of
introduced changes per version.

The new feature aims to create the initial issue for the first version of
the new release and then update that issue for subsequent versions.

1. Create a new issue upon deployment of first RC to staging environment
2. Update the same issue upon deployment of subsequent RCs to staging
  environment
3. Add a comment to the issue to mention and involve (if previously
  unmentioned) relevant stakeholders