Have release-tools take over changelog handling for helm releases.
Previously we have been manualy running a changelog_manager command that
is heavily based on the changelog manger here in this repo. It was
modified to release using the date, rather than the version, as the
chart had weekly milestones.

But now that we have proper release versions, we will only be updating
the changelog from here in the release tools.