Fixes an issue with the Helm chart releasing old security versions
Ensure that the derive chart function does not accept derriving based on
GitLab versions older than what it sees in the current branch.

We can only release old security updates with the chart by passing in
both the chart version, and the gitlab version, as currently we only
check the lastest tag during the derive method.

This was a known limitation when I envisioned the script, but it was not
tested and had not been accounted for properly in the code prior to
this.