Sync release managers group in dev
The `release-managers` group has access to `gitlab-com/chatops` which
will allow the release managers to use chatops.

We also needed to handle errors for users not existing, since there
are still a lot of users that don't exist on ops.gitlab.net. The job
will fail if there are users that don't exist. But all required steps
will be executed.