Option to exclude ci-skip in changelog commit
Previously we were always including ci-skip in the changelog commit

But this is problematic in cases where we want to tag that commit and
have a release pipeline run.

This is not the case for gitlab-ce or gitlab-ee, but our helm charts do
have the changelog generation as the last step before the tag.

Now you can optionally disable the ci-skip option when creating a
changelog manager.