Use full timestamp for upstream branch name
Previously we only used the date in the branch name, which resulted in
multiple branches with the same name being created throughout the day.
In the future we'll want to start fetching pipelines for a specific ref
to ensure that we don't automatically accept an upstream merge request
before a pipeline has been created for it, but we can only do that if
every upstream branch name is actually unique. Otherwise we might be
getting pipelines for merge requests earlier that day.