Add `Issuable#project_id`
The cherry pick summary was receiving a `Gitlab::ObjectifiedHash`
instance for picked MRs, but our a `PreparationMergeRequest` instance
for the summary message. The `project_id` method existed on the former
but not the latter, resulting in a `NoMethodError` when attempting to
post the summary message to the preparation merge request.