Paginate the `user_search` response for RM sync
For usernames with a lot of results, such as `cindy`, the server
responded with more than a single page of results, so we need to loop
through them to find the one we're looking for.

This could have been avoided if we had a sane API that let us fetch a
specific user by their exact username rather than their numeric ID, but
here we are.