Handle release manager schedule not existing
When a schedule definition doesn't exist for the current version, we
were previously returning an empty array, which our API apparently
treats as a bad request, causing a 400 in the client, causing an
exception in release-tools.

We now raise `VersionNotFoundError`, and rescue that in `MonthlyIssue`
in order to return `nil`, which the API likes just fine. Consistency!

See https://sentry.gitlab.net/gitlab/release-tools/issues/662264/