HelmGitlabRelease: send both versions to master
Update HelmGitlabRelease to send both versions to master.
This allows manage_versions script to insert the entry into version
mapping documentation on master as well.

The previous method relied on calling master with only version.
Changes were made in the called script to ignore appVersion changes to
Chart.yml when on the master branch, but add the entry to the version
mapping table in documentation. To do this, we need to send app_version.

See https://gitlab.com/charts/gitlab/merge_requests/704