Automate validating of security merge requests
This adds the ability to validate security merge requests that are
assigned to the release tools bot on dev.gitlab.org. Whenever a merge
request does not meet the requirements it is assigned back to the
author, and a note is created informing the author about what steps they
should take next.