Use case-insensitive matches for tasks
Valid task list completion can be recorded as `[x]` or `[X]`. The former
is more common, as it's what happens when you click the checkbox in the
UI, but the latter is valid too and can come from a manual edit.

The merge request validator should consider both as checked
tasks. Otherwise, it's wrong, and the error messages are very confusing.