Handle missing merge data for QA issues
Sometimes the "merged by" data for an MR is not available, even when
merged. This can happen when (for some reason) data in the
"merge_request_metrics" table for an MR is not present. How and when
this happens is not clear at this time, but it happened for the
following two MRs:

1. https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/27424
2. https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/27400

For the first one I fixed the data manually, but I left the second one
as-is.

When this happens, we now fall back to mentioning the assignee (usually
also the person that merges the MR), or the author if both the merger
and assignee were not present.

Fixes https://gitlab.com/gitlab-org/release-tools/issues/287