Tag auto deploy with Omnibus commit timestamp
We want to tag with something that increments, in order for the package
to be SemVer, but not with something that *always* increments, like the
pipeline ID, even when nothing's changed internally.

This way, when we commit changes to Omnibus, whether through VERSION
file changes or directly in the project itself, the "last commit"
timestamp changes and "updates" the tag version.