Allow ls-remote to return an empty response
`git ls-remote` will return with a zero exit code when it could reach
the remote repository. Than the output depends on whatever the remote
repository dereferenced the reference too. In the case that the
reference isn't known on the server this would mean that there's no
output.

`#ls_remote` is now updated to handle this case.