Remove dead `add_comment` behavior from Qa::Issue
Previously this would add a comment to the QA issue using the
`CommentPresenter`, however this behavior was removed in
dc0047fae6903ee888e4f095d9b1b642fa79271b.

We're retaining the `add_comment` method, which now takes an argument,
because we plan to add comments as the corresponding release transitions
between environments.