Fix changelog generation for Gitaly's CHANGELOG
Gitaly used to prefix the versions with a v in the changelog. The
changelog updater didn't match a string if the version did have that
prefix. Now with this commit will match those lines and correctly insert
new versions without the v prefix.

Also the `#bump_versions` method is removed from the Gitaly release
script, which should work now as expected.