Add a partially-functional SyncRemotesService
This service will publish stable branches and tags for the specified
version for EE, CE, and Omnibus.

Currently the service is heavily gated by feature flags: one to
disable the functionality entirely, and one to disable pushing to any
remotes. After merging, we should enable the former but not the latter,
while we observe `publish` job logs to see how it would behave in the
real world.

Once we're satisfied with its potential behavior, we can enable both
flags, and continue to monitor it in the real world.