Create CE branches based on previous stable branch
This changes the monthly preparation service so that CE stable branches
are created based on the previous stable branch, instead of master. This
way the Git logs don't include commits that are effectively reverted by
syncing EE to CE. Including these commits can be confusing, as one might
think a stable branch contains changes that in reality it does not.