Require the "security" label for security MRs
This label is sometimes missing, resulting in the bot ignoring such
merge requests entirely. This commit changes the validation rules so
that it considers all MRs assigned to the bot, and requires that the
"security" label is present.

Since the "security" label is added by an MR template, the lack of this
label suggests the template may not have been used. This in turn could
lead to there beiong other issues with a security merge request.