Prevent infinite loop in BaseProject
Prior, we were interpolating `self`, which would call `inspect`, which
would call `to_s`, which, in a security release, would call `dev_path`.
When a project didn't define a dev path, such as Deployer, this would
raise the exception again, resulting in the loop.

Now we show the project's default path, which is always defined.