Use the cherry-pick error code returned by API
When the GitLab API fails to cherry-pick, it now includes an
`error_code` field in the JSON response.

I'm hoping that the gitlab gem will just make this field available as a
method on the Error object, as it does for arbitrary fields on other
resources.

If it doesn't, we fall back to the `:failure` code.