Add reason why cherry-pick failed
When a merge request does not have a P1 or P2 label, it will not get
cherry-picked but the bot will not include an explanation why. Instead
of this:

```
@liz.lemon This merge request could not automatically be picked into
`11-4-stable` for `11.4.1`. This requires manual intervention.

Please refer to
[GitLab.com releases](https://about.gitlab.com/handbook/engineering/releases/#gitlabcom-releases-2)

/unlabel ~"Pick into 11.4"
```

We now see:

```
@liz.lemon This merge request could not automatically be picked into
`11-4-stable` for `11.4.1`:

* Merge request does not have P1 or P2 label

This requires manual intervention.

Please refer to
[GitLab.com releases](https://about.gitlab.com/handbook/engineering/releases/#gitlabcom-releases-2)

/unlabel ~"Pick into 11.4"
```