Prevent false positive "successful" commit
When a commit only has docs changes, we run a reduced pipeline to
prevent wasting resources. As a result, a commit might be green while
the code is actually broken from a previous build.

Now a passing build has to have a "full" pipeline in order to be
considered a success.

Because the jobs API might not necessarily be consistent with the order
the jobs are returned in, and because the GitLab CI configuration might
change job names or dependencies, we're relying on the number of jobs in
a successful pipeline as a determining factor.

See https://gitlab.com/gitlab-com/gl-infra/delivery/issues/542