Update Security merger to handle security client
With the security development happening on GitLab.com, we need to merge
security merge requests from different remotes, Security and Dev. To
help with this Security::MergeRequestsMerger was modified to accept
different security clients.

Other modifications included:
* Security merge task was modified to execute Merger with DevClient
and Client if security_remote is enabled. If it's not enabled, then
it's onlyl executed with dev client
* Fix typo on MergeRequestsValidator param comment
* Adjusts Merger specs

Related to https://gitlab.com/gitlab-org/release-tools/issues/344