Rename `latest_dev_green_build_commit` to `latest_successful_on_build`
We're in a class called `Commits` already, so naming a method
`latest_whatever_commit` was redundant. The new name also clarifies that
"green" is "successful", and we're specifically looking for commits that
also exist on Build (dev).

This also better matches the naming of its companion methods, `latest`
and `latest_successful`.