Track deployments using the GitLab API
This adds code to record GitLab and Gitaly deployments using the GitLab
API. This code is a port and refactor of the code currently used in the
deployer to record deployments. Unlike the deployer we split the work in
two different classes, making it easier to use and test the code.

The tracking of deployments can be triggered using a CI job. The
deployer can then trigger this job at some point in the deployment
pipeline.