Notify merge requests deployed to pre
This extends the deployment tracking so that merge request deployed to
the pre environment are updated with a comment. This comment will inform
the merge request participants that the merge request will be included
in the upcoming release, and what that release is. The notified merge
request is also given a new label, making it easier to see which merge
requests have been released.