Makes security patch respond to version.
When security patch tasks were concentrated in a single issue, security
patch object was changed to respond to 'versions', but it also needs to
respond to a single version (the highest one), so it can be backwards
compatible with existent tasks.

* 'version' method was added to retrieve the highest version using
"ReleaseTools::Versions.latest".
* 'milestone_name' method was removed because it's no longer necessary

Related to https://gitlab.com/gitlab-org/release-tools/issues/380